Refactor ServiceUsageSerializer into a reusable utility to calculate usage numbers
