+++ kobo/apps/stripe/tests/test_organization_usage.py
@@ -15,13 +15,13 @@ from kobo.apps.kobo_auth.shortcuts import User
 from kobo.apps.organizations.models import Organization, OrganizationUser
 from kobo.apps.stripe.tests.utils import generate_enterprise_subscription, generate_plan_subscription
 from kobo.apps.trackers.tests.submission_utils import create_mock_assets, add_mock_submissions
-from kpi.tests.api.v2.test_api_service_usage import ServiceUsageAPIBase
+from kpi.tests.test_usage_calculator import BaseUsageCalculatorTestCase
 from kpi.tests.api.v2.test_api_asset_usage import AssetUsageAPITestCase
 from rest_framework import status
 
 
 
-class OrganizationServiceUsageAPIMultiUserTestCase(ServiceUsageAPIBase):
+class OrganizationServiceUsageAPIMultiUserTestCase(BaseUsageCalculatorTestCase):
     """
     Test organization service usage when Stripe is enabled.
 
@@ -149,7 +149,7 @@ class OrganizationServiceUsageAPIMultiUserTestCase(ServiceUsageAPIBase):
             self.expected_file_size() * self.expected_submissions_multi
         )
 
-class OrganizationServiceUsageAPITestCase(ServiceUsageAPIBase):
+class OrganizationServiceUsageAPITestCase(BaseUsageCalculatorTestCase):
     org_id = 'orgAKWMFskafsngf'
 
     @classmethod
